(core) add generic runner and rename module
