+++ lua/snacks-vimtex-output/init.lua
@@ -1,878 +1,2 @@
--- snacks-vimtex-output
--- Render VimTeX compiler output in a floating overlay and surface build status through notifications.
-
-local M = {}
-
-local uv = vim.uv or vim.loop
-
-local default_config = {
-  -- Dimensions for the unfocused mini overlay
-  mini = {
-    width_scale = 0.90,
-    width_min = 48,
-    width_max = 120,
-    height_ratio = 0.10,
-    height_min = 5,
-    height_max = 14,
-    row_anchor = "bottom",
-    row_offset = 5,
-    horizontal_align = 0.55,
-    col_offset = 0,
-  },
-  -- Dimensions for the focused/interactive overlay
-  focus = {
-    width_scale = 0.95,
-    width_min = 70,
-    width_max = 220,
-    height_ratio = 0.6,
-    height_min = 14,
-    height_max = 0,
-    row_anchor = "center",
-    row_offset = -1,
-    horizontal_align = 0.5,
-    col_offset = 0,
-  },
-  -- Extra cursor padding so the overlay never hides the cursor
-  scrolloff_margin = 5,
-  -- Retry logic for auto-opening when the log file is recreated lazily
-  auto_open = {
-    enabled = false,
-    retries = 6,
-    delay = 150,
-  },
-  -- Automatically hide the floating window after a successful build
-  auto_hide = {
-    enabled = true,
-    delay = 3000,
-  },
-  -- Notification configuration and override hook
-  notifications = {
-    enabled = true,
-    title = "VimTeX",
-    persist_failure = true,
-  },
-  notifier = nil,
-  border_highlight = "FloatBorder",
-}
-
-local config = vim.deepcopy(default_config)
-M.config = config
-
-local function assign_config(new_config)
-  for key in pairs(config) do
-    config[key] = nil
-  end
-  for key, value in pairs(new_config) do
-    config[key] = value
-  end
-end
-
-local function auto_open_enabled()
-  if not config.auto_open then
-    return false
-  end
-  return config.auto_open.enabled == true
-end
-
-local state = {
-  win = nil,
-  buf = nil,
-  hide_token = 0,
-  started_at = nil,
-  elapsed = nil,
-  target = nil,
-  timer = nil,
-  last_stat = nil,
-  status = "idle",
-  focused = false,
-  prev_win = nil,
-  border_hl = config.border_highlight,
-  scrolloff_restore = nil,
-  render_retry_token = 0,
-  last_close_reason = nil,
-  augroup = nil,
-  notify = nil,
-  failure_notification = nil,
-  running_notified = false,
-}
-
--- Lazy-load and cache the notification backend (custom, snacks.nvim, or vim.notify fallback).
--- The notifier abstraction allows users to provide their own notification functions while
--- automatically detecting and using snacks.nvim when available, falling back to vim.notify otherwise.
-local function notifier()
-  if state.notify then
-    return state.notify
-  end
-  -- Use custom notifier if provided in config
-  if config.notifier then
-    state.notify = config.notifier
-    return state.notify
-  end
-  -- Default to vim.notify, but upgrade to snacks.notify if available
-  local notify = {
-    info = function(msg, opts)
-      return vim.notify(msg, vim.log.levels.INFO, opts)
-    end,
-    warn = function(msg, opts)
-      return vim.notify(msg, vim.log.levels.WARN, opts)
-    end,
-    error = function(msg, opts)
-      return vim.notify(msg, vim.log.levels.ERROR, opts)
-    end,
-  }
-  -- Detect snacks.nvim and use its richer notification API if present
-  local ok, snacks = pcall(require, "snacks")
-  if ok and snacks.notify then
-    notify.info = function(msg, opts)
-      return snacks.notify.info(msg, opts)
-    end
-    notify.warn = function(msg, opts)
-      return snacks.notify.warn(msg, opts)
-    end
-    notify.error = function(msg, opts)
-      return snacks.notify.error(msg, opts)
-    end
-  end
-  state.notify = notify
-  return notify
-end
-
--- Send a notification at the specified level (info/warn/error) with optional config overrides.
--- Respects config.notifications.enabled and automatically adds title if not provided.
-local function notify(level, msg, opts)
-  local notifications = config.notifications or {}
-  if notifications.enabled == false then
-    return
-  end
-  opts = opts or {}
-  if not opts.title then
-    opts.title = notifications.title or "VimTeX"
-  end
-  local handler = notifier()[level]
-  if handler then
-    local ok, result = pcall(handler, msg, opts)
-    if ok then
-      return result
-    end
-  end
-  return nil
-end
-
-local function clamp(value, min_value, max_value)
-  if min_value then
-    value = math.max(min_value, value)
-  end
-  if max_value and max_value > 0 then
-    value = math.min(max_value, value)
-  end
-  return value
-end
-
-local function current_time()
-  if uv and uv.hrtime then
-    return uv.hrtime()
-  end
-  return nil
-end
-
-local function elapsed_seconds()
-  if state.status == "running" and state.started_at then
-    local now = current_time()
-    if now then
-      return math.max(0, (now - state.started_at) / 1e9)
-    end
-  end
-  return state.elapsed
-end
-
-local function format_duration_suffix()
-  local elapsed = elapsed_seconds()
-  if not elapsed then
-    return ""
-  end
-  return string.format(" (%.1fs)", elapsed)
-end
-
-local status_labels = {
-  running = "Building",
-  success = "Done",
-  failure = "Error",
-}
-
-local function status_title()
-  local label = status_labels[state.status]
-  local elapsed = elapsed_seconds()
-  if label and elapsed then
-    return string.format("VimTeX · %s · %.1fs", label, elapsed)
-  elseif label then
-    return string.format("VimTeX · %s", label)
-  end
-  return "VimTeX"
-end
-
-local function refresh_window_title()
-  if state.win and vim.api.nvim_win_is_valid(state.win) then
-    pcall(vim.api.nvim_win_set_config, state.win, { title = status_title() })
-  end
-end
-
-local function restore_previous_window()
-  local target = state.prev_win
-  state.prev_win = nil
-  if target and vim.api.nvim_win_is_valid(target) then
-    pcall(vim.api.nvim_set_current_win, target)
-    return
-  end
-  for _, win in ipairs(vim.api.nvim_list_wins()) do
-    if win ~= state.win then
-      local ok, cfg = pcall(vim.api.nvim_win_get_config, win)
-      if ok and cfg.relative == "" then
-        pcall(vim.api.nvim_set_current_win, win)
-        return
-      end
-    end
-  end
-end
-
-local function stop_polling()
-  if state.timer then
-    state.timer:stop()
-    if state.timer.close then
-      state.timer:close()
-    end
-    state.timer = nil
-  end
-end
-
-local function close_window(opts)
-  opts = opts or {}
-  stop_polling()
-  if state.win and vim.api.nvim_win_is_valid(state.win) then
-    vim.api.nvim_win_close(state.win, true)
-  end
-  state.win = nil
-  state.focused = false
-  state.render_retry_token = state.render_retry_token + 1
-  state.last_close_reason = opts.reason
-  if state.scrolloff_restore ~= nil then
-    vim.o.scrolloff = state.scrolloff_restore
-    state.scrolloff_restore = nil
-  end
-  restore_previous_window()
-end
-
-local function schedule_hide(delay)
-  if not config.auto_hide or config.auto_hide.enabled == false then
-    return
-  end
-  state.hide_token = state.hide_token + 1
-  local token = state.hide_token
-  local hide_delay = delay or config.auto_hide.delay or 2000
-  vim.defer_fn(function()
-    if state.hide_token == token and not state.focused then
-      close_window({ reason = "auto_success" })
-    end
-  end, hide_delay)
-end
-
-local function normalized_path(path)
-  if not path or path == "" then
-    return nil
-  end
-  return vim.fn.fnamemodify(path, ":p")
-end
-
-local function compiler_output_path(bufnr)
-  local vt
-  if bufnr and vim.api.nvim_buf_is_valid(bufnr) then
-    local ok, buf_vt = pcall(vim.api.nvim_buf_get_var, bufnr, "vimtex")
-    if ok then
-      vt = buf_vt
-    end
-  end
-  vt = vt or vim.b.vimtex
-  if not vt or not vt.compiler or not vt.compiler.output then
-    return nil
-  end
-  return normalized_path(vt.compiler.output)
-end
-
-local function set_target(path)
-  local normalized = normalized_path(path)
-  if not normalized then
-    return nil
-  end
-  if state.target ~= normalized then
-    state.target = normalized
-    state.last_stat = nil
-  end
-  return normalized
-end
-
-local function resolve_target(path, source_buf)
-  local normalized = set_target(path)
-  if normalized then
-    return normalized
-  end
-  local from_source = compiler_output_path(source_buf)
-  if from_source then
-    return set_target(from_source)
-  end
-  if state.target then
-    return state.target
-  end
-  local current = compiler_output_path()
-  if current then
-    return set_target(current)
-  end
-  return nil
-end
-
-local function ensure_output_buffer()
-  if state.buf and vim.api.nvim_buf_is_valid(state.buf) then
-    return state.buf
-  end
-  local buf = vim.api.nvim_create_buf(false, true)
-  vim.api.nvim_set_option_value("bufhidden", "hide", { buf = buf })
-  vim.api.nvim_set_option_value("buftype", "nofile", { buf = buf })
-  vim.api.nvim_set_option_value("swapfile", false, { buf = buf })
-  vim.api.nvim_set_option_value("modifiable", false, { buf = buf })
-  vim.api.nvim_set_option_value("filetype", "vimtexoutput", { buf = buf })
-  state.buf = buf
-  return buf
-end
-
-local function scroll_to_bottom()
-  if state.focused or not state.win or not vim.api.nvim_win_is_valid(state.win) then
-    return
-  end
-  if not state.buf or not vim.api.nvim_buf_is_valid(state.buf) then
-    return
-  end
-  local line_count = vim.api.nvim_buf_line_count(state.buf)
-  if line_count == 0 then
-    return
-  end
-  pcall(vim.api.nvim_win_set_cursor, state.win, { line_count, 0 })
-end
-
-local function update_buffer_from_file(force)
-  local target = state.target
-  if not target then
-    return false
-  end
-  local buf = ensure_output_buffer()
-  if not buf then
-    return false
-  end
-  local stat = uv and target and uv.fs_stat(target) or nil
-  if not stat then
-    return false
-  end
-  local changed = force
-    or not state.last_stat
-    or stat.mtime.sec ~= state.last_stat.mtime.sec
-    or stat.mtime.nsec ~= state.last_stat.mtime.nsec
-    or stat.size ~= state.last_stat.size
-  if not changed then
-    return true
-  end
-  local ok, lines = pcall(vim.fn.readfile, target)
-  if not ok then
-    return false
-  end
-  if #lines == 0 then
-    lines = { "" }
-  end
-  local mod = vim.api.nvim_get_option_value("modifiable", { buf = buf })
-  if not mod then
-    vim.api.nvim_set_option_value("modifiable", true, { buf = buf })
-  end
-  vim.api.nvim_buf_set_lines(buf, 0, -1, false, lines)
-  vim.api.nvim_set_option_value("modified", false, { buf = buf })
-  if not mod then
-    vim.api.nvim_set_option_value("modifiable", false, { buf = buf })
-  end
-  state.last_stat = stat
-  scroll_to_bottom()
-  return true
-end
-
--- Keep a lightweight polling loop running while the overlay is visible so the
--- buffer stays in sync with latexmk's stdout and the title timer keeps
--- updating. The timer exits automatically when the floating window disappears
--- (either because the user hid it or VimTeX stopped the build) and spins up
--- again the next time `render_window` requests live updates.
-local function start_polling()
-  if state.timer or not uv then
-    return
-  end
-  local timer = uv.new_timer()
-  state.timer = timer
-  timer:start(0, 250, function()
-    vim.schedule(function()
-      local win_open = state.win and vim.api.nvim_win_is_valid(state.win)
-      if not win_open then
-        stop_polling()
-        return
-      end
-      if not state.target then
-        stop_polling()
-        return
-      end
-      local buf = ensure_output_buffer()
-      if not buf or not vim.api.nvim_buf_is_valid(buf) then
-        stop_polling()
-        return
-      end
-      update_buffer_from_file(false)
-      refresh_window_title()
-    end)
-  end)
-end
-
-local function ensure_output_ready(opts)
-  opts = opts or {}
-  local target = resolve_target(opts.target, opts.source_buf)
-  if not target then
-    if not opts.quiet then
-      notify("warn", "VimTeX does not expose a compiler output log")
-    end
-    return nil
-  end
-  local buf = ensure_output_buffer()
-  if not buf then
-    return nil
-  end
-  update_buffer_from_file(true)
-  return buf
-end
-
--- Calculate window dimensions and position based on current terminal size and layout config.
--- @param focused boolean: true for focus mode (large interactive window), false for mini mode (compact overlay)
--- @return width, height, row, col: dimensions and position for nvim_open_win
-local function window_dimensions(focused)
-  local layout = focused and config.focus or config.mini
-  local cols = vim.o.columns
-  local lines = vim.o.lines
-
-  -- Calculate width as a percentage of terminal width, then clamp to min/max bounds
-  local width = math.floor(cols * (layout.width_scale or 0.75))
-  width = clamp(width, layout.width_min or 40, layout.width_max or cols)
-
-  -- Calculate height as a percentage of terminal height, then clamp to min/max bounds
-  local height = math.floor(lines * (layout.height_ratio or 0.3))
-  height = clamp(height, layout.height_min or 6, layout.height_max or math.max(2, lines - 2))
-
-  -- Horizontal positioning: align is a 0-1 value where 0=left, 0.5=center, 1=right
-  -- Then apply col_offset as a fine-tuning adjustment
-  local align = clamp(layout.horizontal_align or 0.5, 0, 1)
-  local col = math.floor((cols - width) * align + (layout.col_offset or 0))
-  col = clamp(col, 0, math.max(0, cols - width))
-
-  -- Vertical positioning: anchor determines where row_offset is measured from
-  -- top: offset from top edge; center: offset from vertical center; bottom: offset from bottom edge
-  local anchor = layout.row_anchor or "bottom"
-  local row_offset = layout.row_offset or 0
-  local row
-  if anchor == "top" then
-    row = row_offset
-  elseif anchor == "center" then
-    row = math.floor((lines - height) / 2) + row_offset
-  else
-    row = lines - height - row_offset
-  end
-  row = clamp(row, 0, math.max(0, lines - height))
-
-  return width, height, row, col
-end
-
--- Temporarily increase scrolloff in mini mode to prevent the overlay from covering the cursor.
--- In focus mode, the user controls cursor position, so we restore the original scrolloff.
--- This ensures the cursor always has enough padding when typing while the mini overlay is visible.
-local function apply_cursor_guard(height, focused)
-  if focused then
-    -- Restore original scrolloff when entering focus mode
-    if state.scrolloff_restore ~= nil then
-      vim.o.scrolloff = state.scrolloff_restore
-      state.scrolloff_restore = nil
-    end
-    return
-  end
-  -- In mini mode, ensure scrolloff is large enough to keep cursor visible above the overlay
-  local margin = config.scrolloff_margin or 1
-  local needed = math.max(0, height + margin)
-  if needed == 0 then
-    return
-  end
-  -- Save original scrolloff only once (when first showing mini overlay)
-  if state.scrolloff_restore == nil then
-    state.scrolloff_restore = vim.o.scrolloff
-  end
-  -- Increase scrolloff if current value is too small
-  if vim.o.scrolloff < needed then
-    vim.o.scrolloff = needed
-  end
-end
-
-local function render_window(opts)
-  opts = opts or {}
-  local should_poll = opts.poll
-  if should_poll == nil then
-    should_poll = state.status == "running"
-  end
-  local buf = ensure_output_ready({
-    target = opts.target,
-    source_buf = opts.source_buf,
-    quiet = opts.quiet,
-  })
-  if not buf then
-    return nil
-  end
-
-  local desired_focus = opts.focus
-  if desired_focus == nil then
-    desired_focus = state.focused
-  end
-
-  local entering_focus = desired_focus and not state.focused
-  local leaving_focus = state.focused and not desired_focus
-
-  if entering_focus then
-    local current = vim.api.nvim_get_current_win()
-    if current ~= state.win then
-      state.prev_win = current
-    end
-  end
-
-  local width, height, row, col = window_dimensions(desired_focus)
-  local config = {
-    relative = "editor",
-    width = width,
-    height = height,
-    row = row,
-    col = col,
-    border = "rounded",
-    focusable = desired_focus,
-    style = "minimal",
-    title = opts.title or status_title(),
-  }
-
-  local border_hl = opts.border_hl or state.border_hl or config.border_highlight or "FloatBorder"
-  state.border_hl = border_hl
-
-  local win_exists = state.win and vim.api.nvim_win_is_valid(state.win)
-  if win_exists then
-    vim.api.nvim_win_set_buf(state.win, buf)
-    vim.api.nvim_win_set_config(state.win, config)
-  else
-    local open_config = vim.deepcopy(config)
-    open_config.noautocmd = true
-    state.win = vim.api.nvim_open_win(buf, desired_focus, open_config)
-  end
-
-  vim.api.nvim_set_option_value("wrap", false, { win = state.win })
-  vim.api.nvim_set_option_value("cursorline", desired_focus, { win = state.win })
-  vim.api.nvim_set_option_value(
-    "winhl",
-    string.format("Normal:NormalFloat,FloatBorder:%s", border_hl),
-    { win = state.win }
-  )
-
-  if desired_focus then
-    if vim.api.nvim_get_current_win() ~= state.win then
-      vim.api.nvim_set_current_win(state.win)
-    end
-  elseif leaving_focus then
-    restore_previous_window()
-  end
-
-  apply_cursor_guard(height, desired_focus)
-  state.focused = desired_focus
-  refresh_window_title()
-  if not desired_focus then
-    scroll_to_bottom()
-  end
-  if should_poll then
-    start_polling()
-  end
-  state.last_close_reason = nil
-  return state.win
-end
-
--- Attempt to render the window with retry logic to handle race conditions.
--- VimTeX sometimes recreates the log file during compilation startup, causing the initial
--- render to fail because the file doesn't exist yet. This function retries after a delay.
--- Uses token-based cancellation to abort retries if the window is closed or compilation stops.
-local function render_with_retry(opts)
-  opts = opts or {}
-  local attempts = opts.retry_count or 0
-  local delay = opts.retry_delay or 100
-  -- Strip retry parameters before passing to render_window
-  local render_opts = vim.deepcopy(opts)
-  render_opts.retry_count = nil
-  render_opts.retry_delay = nil
-  -- Try rendering immediately first
-  if render_window(render_opts) then
-    return
-  end
-  -- If no retries configured or render succeeded, we're done
-  if attempts <= 0 then
-    return
-  end
-  -- Increment token to cancel any previous retry loops
-  state.render_retry_token = state.render_retry_token + 1
-  local token = state.render_retry_token
-  -- Recursive retry function with token-based cancellation
-  local function retry(remaining)
-    if remaining <= 0 then
-      return
-    end
-    vim.defer_fn(function()
-      -- Abort if token changed (window closed) or compilation stopped
-      if state.render_retry_token ~= token or state.status ~= "running" then
-        return
-      end
-      if render_window(render_opts) then
-        -- Success! Invalidate this retry chain
-        state.render_retry_token = state.render_retry_token + 1
-        return
-      end
-      -- Still failing, try again
-      retry(remaining - 1)
-    end, delay)
-  end
-  retry(attempts)
-end
-
-local function open_overlay(opts)
-  opts = opts or {}
-  if opts.focus == nil then
-    opts.focus = false
-  end
-  opts.source_buf = opts.source_buf or vim.api.nvim_get_current_buf()
-  if state.status == "running" then
-    opts.poll = true
-    opts.retry_count = opts.retry_count or (config.auto_open and config.auto_open.retries) or 0
-    opts.retry_delay = opts.retry_delay or (config.auto_open and config.auto_open.delay) or 120
-    render_with_retry(opts)
-  else
-    render_window(opts)
-  end
-end
-
-function M.show()
-  state.hide_token = state.hide_token + 1
-  state.render_retry_token = state.render_retry_token + 1
-  open_overlay()
-end
-
-function M.hide()
-  state.hide_token = state.hide_token + 1
-  close_window({ reason = "manual" })
-end
-
-function M.toggle()
-  if state.win and vim.api.nvim_win_is_valid(state.win) then
-    M.hide()
-  else
-    M.show()
-  end
-end
-
-function M.toggle_focus()
-  state.hide_token = state.hide_token + 1
-  state.render_retry_token = state.render_retry_token + 1
-  if not state.win or not vim.api.nvim_win_is_valid(state.win) then
-    open_overlay()
-    return
-  end
-  render_window({ focus = not state.focused })
-end
-
--- Extract VimTeX event context to determine which buffer and log file to track.
--- VimTeX events include a buffer reference; we use it to find the compiler output path.
-local function event_context(event)
-  local buf = event and event.buf and event.buf ~= 0 and event.buf or nil
-  local target = compiler_output_path(buf)
-  if target then
-    set_target(target)
-  end
-  return { target = target or state.target, source_buf = buf }
-end
-
--- Handle VimTeX compilation start events (VimtexEventCompiling, VimtexEventCompileStarted).
--- Transitions state to "running", starts the timer if needed, and shows the overlay if configured.
-local function on_compile_started(event)
-  local ctx = event_context(event)
-  local already_running = state.status == "running"
-  -- Only reset timer on the first compile event (not subsequent recompiles)
-  if not already_running then
-    state.started_at = current_time()
-    state.elapsed = nil
-  end
-  state.status = "running"
-  state.border_hl = config.border_highlight or "FloatBorder"
-  state.hide_token = state.hide_token + 1 -- Cancel any pending auto-hide
-  local window_exists = state.win and vim.api.nvim_win_is_valid(state.win)
-  -- Show overlay if auto-open is enabled OR if the user already opened it manually
-  if auto_open_enabled() or window_exists then
-    render_with_retry({
-      target = ctx.target,
-      source_buf = ctx.source_buf,
-      poll = true,
-      focus = state.focused,
-      quiet = true,
-      retry_count = (config.auto_open and config.auto_open.retries) or 0,
-      retry_delay = (config.auto_open and config.auto_open.delay) or 120,
-    })
-  end
-  if not state.running_notified then
-    notify("info", "LaTeX build started…")
-    state.running_notified = true
-  end
-end
-
--- Handle VimTeX compilation success event (VimtexEventCompileSuccess).
--- Stops polling, transitions to "success" state, updates border to green, and schedules auto-hide.
--- Replaces any previous failure notification with a success message.
-local function on_compile_succeeded(event)
-  local ctx = event_context(event)
-  local elapsed = elapsed_seconds()
-  state.elapsed = elapsed
-  state.started_at = nil
-  state.status = "success"
-  state.running_notified = false
-  stop_polling()
-  local window_exists = state.win and vim.api.nvim_win_is_valid(state.win)
-  if auto_open_enabled() or window_exists then
-    render_window({
-      border_hl = "DiagnosticOk", -- Green border for success
-      target = ctx.target,
-      source_buf = ctx.source_buf,
-      focus = state.focused,
-    })
-    schedule_hide(config.auto_hide and config.auto_hide.delay)
-  end
-  -- Replace previous failure notification if one exists
-  local notify_opts
-  if state.failure_notification then
-    notify_opts = { replace = state.failure_notification }
-  end
-  notify("info", "LaTeX build finished" .. format_duration_suffix(), notify_opts)
-  state.failure_notification = nil
-end
-
--- Handle VimTeX compilation failure event (VimtexEventCompileFailed).
--- Stops polling, transitions to "failure" state, updates border to red, and shows persistent notification.
--- The overlay stays visible (auto-hide is cancelled) so the user can inspect the error log.
-local function on_compile_failed(event)
-  local ctx = event_context(event)
-  local elapsed = elapsed_seconds()
-  state.elapsed = elapsed
-  state.started_at = nil
-  state.status = "failure"
-  state.running_notified = false
-  stop_polling()
-  state.hide_token = state.hide_token + 1 -- Cancel any pending auto-hide
-  local window_exists = state.win and vim.api.nvim_win_is_valid(state.win)
-  if auto_open_enabled() or window_exists then
-    render_window({
-      border_hl = "DiagnosticError", -- Red border for failure
-      target = ctx.target,
-      source_buf = ctx.source_buf,
-      focus = state.focused,
-    })
-  end
-  -- Show persistent failure notification (if configured) so it stays visible until next success
-  local notifications = config.notifications or {}
-  local failure_opts = {
-    replace = state.failure_notification,
-  }
-  if notifications.persist_failure ~= false then
-    failure_opts.timeout = false -- Don't auto-dismiss
-    failure_opts.keep = true -- Keep in notification history
-  end
-  state.failure_notification =
-    notify("error", "LaTeX build failed" .. format_duration_suffix(), failure_opts)
-end
-
-local function on_compile_stopped()
-  local was_running = state.status == "running"
-  state.status = was_running and "idle" or state.status
-  state.started_at = nil
-  state.elapsed = was_running and nil or state.elapsed
-  state.running_notified = false
-  stop_polling()
-  if was_running or state.win then
-    close_window({ reason = "stopped" })
-  else
-    refresh_window_title()
-  end
-end
-
-local function setup_autocmds()
-  if state.augroup then
-    pcall(vim.api.nvim_del_augroup_by_id, state.augroup)
-  end
-  local group = vim.api.nvim_create_augroup("snacks_vimtex_output", { clear = true })
-  state.augroup = group
-  vim.api.nvim_create_autocmd("User", {
-    group = group,
-    pattern = "VimtexEventCompiling",
-    callback = on_compile_started,
-  })
-  vim.api.nvim_create_autocmd("User", {
-    group = group,
-    pattern = "VimtexEventCompileStarted",
-    callback = on_compile_started,
-  })
-  vim.api.nvim_create_autocmd("User", {
-    group = group,
-    pattern = "VimtexEventCompileSuccess",
-    callback = on_compile_succeeded,
-  })
-  vim.api.nvim_create_autocmd("User", {
-    group = group,
-    pattern = "VimtexEventCompileFailed",
-    callback = on_compile_failed,
-  })
-  vim.api.nvim_create_autocmd("User", {
-    group = group,
-    pattern = "VimtexEventCompileStopped",
-    callback = on_compile_stopped,
-  })
-end
-
-local function setup_commands()
-  local commands = {
-    VimtexOutputShow = M.show,
-    VimtexOutputHide = M.hide,
-    VimtexOutputToggle = M.toggle,
-    VimtexOutputToggleFocus = M.toggle_focus,
-  }
-  for name in pairs(commands) do
-    pcall(vim.api.nvim_del_user_command, name)
-  end
-  for name, fn in pairs(commands) do
-    vim.api.nvim_create_user_command(name, function()
-      fn()
-    end, {})
-  end
-end
-
-function M.setup(opts)
-  local merged = vim.tbl_deep_extend("force", vim.deepcopy(default_config), opts or {})
-  assign_config(merged)
-  state.border_hl = config.border_highlight or "FloatBorder"
-  state.notify = nil
-  state.failure_notification = nil
-  state.running_notified = false
-  setup_autocmds()
-  setup_commands()
-  return config
-end
-
-return M
-
--- vim: ts=2 sts=2 sw=2 et
+-- Compatibility shim so existing configs using `require("snacks-vimtex-output")` keep working.
+return require("output-panel")
